[Reka and Fix] move gpt eval to process_results stage. (#108)
* Refactor activitynetqa_generation.yaml for improved metric aggregation

* Refactor JSON dumping in cli_evaluate_single for non-ASCII characters

* Refactor video_detail_description/utils.py for improved cache directory handling

* Refactor gpt_eval function to handle non-ASCII characters in data_dict

* Refactor pywsd import error message for lmms-eval nextqa module

* Refactor lmms_eval/models/llava_vid.py for consistency in variable naming

* Refactor activitynetqa default template YAML for improved prompt formatting

* Refactor lmms_eval/tasks/activitynetqa/utils.py for consistent capitalization of questions

* Refactor llava_vid.py for improved scaling factor calculation

+++ lmms_eval/tasks/videochatgpt/utils.py
@@ -404,7 +404,7 @@ def parse_score(review):
         # Convert the string representation of a dictionary to an actual dictionary
         review_dict = ast.literal_eval(review)
         score = review_dict.get("score", 0)
-        return float(score)
+        return int(score)
     except SyntaxError as e:
         eval_logger.error(f"Syntax error parsing the review string: {e}. Review content: {review}")
         return 0
